tidyup stories
