+++ src/components/Chart/chartData.ts
@@ -1,0 +1,97 @@
+export const barData = [
+  { month: "Jan", value: 65 },
+  { month: "Feb", value: 75 },
+  { month: "Mar", value: 90 },
+  { month: "Apr", value: 85 },
+  { month: "May", value: 95 },
+];
+
+export const multiSeriesData = [
+  { month: "Jan", sales: 65, profit: 45, cost: 20 },
+  { month: "Feb", sales: 75, profit: 50, cost: 25 },
+  { month: "Mar", sales: 90, profit: 65, cost: 25 },
+  { month: "Apr", sales: 85, profit: 55, cost: 30 },
+  { month: "May", sales: 95, profit: 70, cost: 25 },
+];
+
+export const lineData = [
+  {
+    id: "sales",
+    data: [
+      { x: "Jan", y: 65 },
+      { x: "Feb", y: 75 },
+      { x: "Mar", y: 90 },
+      { x: "Apr", y: 85 },
+      { x: "May", y: 95 },
+    ],
+  },
+  {
+    id: "profit",
+    data: [
+      { x: "Jan", y: 45 },
+      { x: "Feb", y: 50 },
+      { x: "Mar", y: 65 },
+      { x: "Apr", y: 55 },
+      { x: "May", y: 70 },
+    ],
+  },
+];
+
+export const pieData = [
+  { id: "JavaScript", value: 35 },
+  { id: "TypeScript", value: 25 },
+  { id: "Python", value: 20 },
+  { id: "Go", value: 15 },
+  { id: "Rust", value: 5 },
+];
+
+export const stackedData = [
+  {
+    category: "凡例1",
+    segment1: 20,
+    segment2: 20,
+    segment3: 20,
+    segment4: 20,
+    segment5: 20,
+  },
+  {
+    category: "凡例2",
+    segment1: 20,
+    segment2: 20,
+    segment3: 20,
+    segment4: 20,
+    segment5: 0,
+  },
+  {
+    category: "凡例3",
+    segment1: 20,
+    segment2: 20,
+    segment3: 20,
+    segment4: 20,
+    segment5: 20,
+  },
+  {
+    category: "凡例4",
+    segment1: 20,
+    segment2: 20,
+    segment3: 20,
+    segment4: 20,
+    segment5: 0,
+  },
+  {
+    category: "凡例5",
+    segment1: 20,
+    segment2: 20,
+    segment3: 20,
+    segment4: 20,
+    segment5: 0,
+  },
+  {
+    category: "凡例6",
+    segment1: 20,
+    segment2: 20,
+    segment3: 20,
+    segment4: 20,
+    segment5: 0,
+  },
+];
